Fetch questions query only when leaf node is expanded
